// src/app/models/clienteDomicilio.ts
export interface ClienteDomicilio {
    id_evt_Cliente_Domicilio: number;
    id_evt_Cliente: number;
    Domicilio_Cliente_Domicilio: string;
    Localidad_Cliente_Domicilio: string;
    CPostal_Cliente_Domicilio: string;
    Provincia_Cliente_Domicilio: string;
    Pais_Cliente_Domicilio: string;
    Mail_Despacho: string;
    Telefono_Cliente_Domicilio: string;
    Domicilio_Central_Cliente: string;
}